+++ test/ensure_build_test.dart
@@ -1,3 +1,4 @@
+@Tags(['build'])
 import 'package:build_verify/build_verify.dart';
 import 'package:test/test.dart';